\name{ICOTS2014-package}
\alias{ICOTS2014-package}
\alias{ICOTS2014}
\docType{package}
\title{Distribution of files for the Project MOSAIC workshop at ICOTS-2014. 
}
\description{
Mainly a way to distribute RMarkdown template files for RStudio.
}
\details{
\tabular{ll}{
Package: \tab ICOTS2014\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-12\cr
License: \tab What license is it under?\cr
}
Install this package.  Then, when opening an Rmd document, select the appropriate file from the listing under "From Template."
}
\author{Substantive contents: Nick Horton, Randall Pruim, and Daniel Kaplan.  The package itself: Daniel Kaplan.
}
\references{
See the Project MOSAIC books distributed at the workshop.
}
\keyword{ package }
\seealso{
\code{\link[mosaic:mosaic-package]{mosaic}}
}
\examples{
}
